requirements.txt
numpy
Pillow
tensorflow-gpu==2.2
tabulate
tqdm
